Revert "For #12171 - Add a PocketStory supertype for all Pocket stories types"
This reverts commit 10d016442c6cdf24a97085c0aae6e75aaced782c.
